Finished filters.rs
Added enum for custom errors, wrote object_identifier_filter,
array_index, array_slive, array_iterator, and pipe for various ways
to filter to input. Wrote another enum FilterResult for handling
the two different types of output, a single Value or an iterator,
and implemented PartialEq and Debug for it. Wrote the main
filter_input function which runs all the filters according to the
needle provided by the user and also the add/length/del functions.
Added unit tests to test all the functions